simplify files
