+++ tests/ArrayHelpersTest.php
@@ -4,13 +4,8 @@ namespace Test;
 
 use PHPUnit\Framework\TestCase;
 
-class HelpersTest extends TestCase
+class ArrayHelpersTest extends TestCase
 {
-    public function test_str_contains()
-    {
-        $this->assertTrue(str_contains('Laravel', 'Lara'));
-    }
-
     /**
      * @dataProvider arrayFirstData
      *
